+++ pub/flutter_unrouter/lib/src/core/route_definition.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:collection';
 
 import 'package:flutter/widgets.dart';
 import 'package:unrouter/unrouter.dart'